Partial fix for IVY-1649
The namespaces of dependencies' extra attributes were forgotten before
writing them to the cache when a module descriptor was created from
scratch. With this change, namespaced attributes are correctly
transferred to the generated module descriptor, but the xmlns
declarations still have to be specified locally on each dependency
element (see IVY-1650).